+++ include/IWindowManager.h
@@ -5,7 +5,13 @@
  * Copyright (c) 2020, 2021 Roman Divotkey, Nora Loimayr. All rights reserved.
  */
 
+#pragma once
+
+// C++ Standard Libryry includes
 #include <string>
+
+// Local includes
+#include "Color.h"
 
 namespace astu {